pyproject.toml
[project]
name = "datasense"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "fastapi>=0.135.2",
    "google-genai>=1.68.0",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.2.1",
    "langchain-text-splitters>=1.1.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.11",
    "pypdf>=6.9.2",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "rank-bm25>=0.2.2",
    "reportlab>=4.4.10",
    "uvicorn>=0.42.0",
]
